requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
Pillow>=9.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
hf_transfer>=0.1.0
